+++ NEWS.md
@@ -1,6 +1,6 @@
 # FSA 0.7.7 ongoing
-* `ageBias()`: Modified.  Changed default for `pch.mean=` to 95 (from 175).
-* `agePrecision()`: Modified.  Added `trunc.diff=`.
+* `ageBias()`: Modified.  Changed default for `pch.mean=` to 95 (from 175).  If `what=` has only one item, then results will now be invisibly returned so that results can be saved to an object.
+* `agePrecision()`: Modified.  Added `trunc.diff=`.  If `what=` has only one item, then results will now be invisibly returned so that results can be saved to an object.
 * `mapvalues()`: Modified.  Corrected to export properly.
 * `removal()`: Modified.  Minor edits to labels if `verbose=TRUE`.  Added some more tests.